refactor: Standardize imports and enhance error handling in components
This commit refines the import statements across multiple components for consistency and clarity. Key changes include:

- Updated import paths for `EmptyState` and `ConfirmDialog` to streamline usage from the `ui` module.
- Enhanced error handling in the `useAppContext` hook to provide clearer guidance on usage context.
- Improved the `SongDetailRoute` component to handle missing songs more gracefully, avoiding unnecessary toast notifications on every render.

These updates contribute to a cleaner codebase and improved user experience by ensuring better error management and consistent component imports.
