+++ src/components/Navigation.tsx
@@ -79,7 +79,7 @@ export const Navigation: React.FC<NavigationProps> = ({
         />
       )}
 
-      <nav className="flex-1 px-3 py-6 space-y-2">
+      <nav className="flex-1 px-3 py-6 space-y-2" aria-label="Main navigation">
         {navItems.map(item => (
           <button
             key={item.id}
